+++ src/test/java/janggi/domain/piece/HorseTest.java
@@ -1,11 +1,8 @@
-package janggi.piece;
+package janggi.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import janggi.domain.board.Position;
-import janggi.domain.piece.Horse;
-import janggi.domain.piece.Piece;
-import janggi.domain.piece.Side;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
