+++ src/test/java/janggi/domain/piece/SoldierTest.java
@@ -1,10 +1,8 @@
-package janggi.piece;
+package janggi.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import janggi.domain.board.Position;
-import janggi.domain.piece.Side;
-import janggi.domain.piece.Soldier;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
