+++ src/test/java/janggi/domain/piece/TankTest.java
@@ -1,11 +1,8 @@
-package janggi.piece;
+package janggi.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import janggi.domain.board.Position;
-import janggi.domain.piece.Piece;
-import janggi.domain.piece.Side;
-import janggi.domain.piece.Tank;
 import java.util.Map;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
